requirements.txt
python-telegram-bot
python-dotenv
apscheduler
requests
google-generativeai